quick_translator/requirements.txt
google-cloud-translate==3.0.2
flask==1.1.2
gunicorn==20.0.4